# backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Smart Campus API",
    "version": "2.0.0",
    "description": "API de la plateforme Smart Campus (ESP/UCAD). Couvre l'authentification MFA, le contrôle d'accès multi-niveaux, les paiements (portefeuille, services, scolarité), la gestion des présences et l'interconnexion avec le SI de l'ESP.\n\nSécurité : JWT (Bearer) pour les routes d'administration ; carte NFC + PIN pour les paiements ; signature HMAC pour le webhook ESP."
  },
  "servers": [
    { "url": "https://smart-campus-jgn5.onrender.com/api", "description": "Production (Render)" },
    { "url": "http://localhost:5000/api", "description": "Développement local" }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" },
      "hmacSignature": { "type": "apiKey", "in": "header", "name": "X-ESP-Signature", "description": "HMAC-SHA256(corps_brut, ESP_WEBHOOK_SECRET) en hexadécimal" },
      "readerToken": { "type": "apiKey", "in": "header", "name": "X-Reader-Token", "description": "Jeton du lecteur de confiance d'un espace" }
    },
...
  },
  "paths": {
...
    "/scholarship/stats": {
      "get": {
        "tags": ["Scolarité (frais)"],
        "summary": "Statistiques de collecte (admin/finance)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "academicYear", "in": "query", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Synthèse de collecte" } }
      }
    },
...
    "/attendance/alerts": {
      "get": {
        "tags": ["Présences"],
        "summary": "Étudiants sous le seuil de 75%",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "courseId", "in": "query", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Liste des alertes" } }
      }
    },
    "/courses": {
      "get": {
        "tags": ["Cours"],
        "summary": "Lister les cours (admin/enseignant)",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Liste des cours" } }
      },
      "post": {
        "tags": ["Cours"],
        "summary": "Créer un cours (admin)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["code", "name", "academicYear"], "properties": { "code": { "type": "string" }, "name": { "type": "string" }, "department": { "type": "string" }, "level": { "type": "string" }, "instructorId": { "type": "string" }, "academicYear": { "type": "string" }, "schedule": { "type": "array", "items": { "type": "object" } } } } } } },
        "responses": { "201": { "description": "Cours créé" } }
      }
    },
...
  }
}